Add files via upload
And updated SVG of the top layer, optimized to avoid as many overlapping cuts as possible.

Also included are the .NGC files for creating samples on the laser cutter.  Run the etch file first at full power, followed by the cut file.  They have the same origins and should overlap when loading in LinuxCNC.

The etch file has been updated to use M4 (speed-independent power) so that it will render properly.